Externalize regexGroupToArray() function
