+++ src/util/prolog/prologResult.ts
@@ -1,3 +1,5 @@
+import {regexGroupToArray} from "../helperFunctions";
+
 // First Group will be the variable value. Doesn't fetch the variable names
 const variableAnswerPattern: string = " = (\\w*|\\[.*\\])?(,| ;|.)";
 
@@ -52,7 +54,7 @@ export class PrologResult {
 	private getResultsFor(variable: string): string[] {
 		// Group 1 will be the content of the variable independent of the variable length.
 		const pattern = new RegExp(variable + variableAnswerPattern, "g");
-		return this.regexGroupToArray(this.rawResults.toString(), pattern, 1);
+		return regexGroupToArray(this.rawResults.toString(), pattern, 1);
 	}
 
 	private rawResultsToInnerArray(source: string): string[] {
@@ -62,31 +64,20 @@ export class PrologResult {
 		if (booleanAnswer !== null) { return [booleanAnswer[1]]; }
 
 		const pattern = new RegExp(variableAnswerPattern, "g");
-		return this.regexGroupToArray(source, pattern, 1);
+		return regexGroupToArray(source, pattern, 1);
 	}
 
 	private getVariables(): Set<string> {
-		const rawVariables: string[] = this.regexGroupToArray(this.rawResults.toString(), variableNameRegExp, 2);
+		const rawVariables: string[] = regexGroupToArray(this.rawResults.toString(), variableNameRegExp, 2);
 
 		return new Set(rawVariables);
 	}
 
 	private getBooleans(): boolean[] {
 		const pattern = /(true|false)(,| ;|.)/g;
-		const results: string[] = this.regexGroupToArray(this.rawResults.toString(), pattern, 1);
+		const results: string[] = regexGroupToArray(this.rawResults.toString(), pattern, 1);
 		const booleanResults: boolean[] = results.map((e: string) => e === "true");
 
 		return booleanResults;
-	}
-
-	private regexGroupToArray(basis: string, pattern: RegExp, group: number): string[] {
-		const groupContent: string[] = [];
-		let matchGroups: string[] | null;
-
-		while ((matchGroups = pattern.exec(basis)) !== null) {
-			groupContent.push(matchGroups[group]);
-		}
-
-		return groupContent;
 	}
 }
